phase-1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
anthropic==0.7.8
pyjwt==2.8.0
python-multipart==0.0.6
jinja2==3.1.2
